requirements.txt
Django>=5.2
djangorestframework>=3.16
requests>=2.32
djangorestframework-simplejwt>=5.5.0
pandas>=2.2.3
